requirements.txt
pytorch3d
pybullet
trimesh